+++ app/src/androidTest/java/woowacourse/movie/HomeFragmentTest.kt
@@ -16,7 +16,7 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import woowacourse.movie.activity.MainActivity
-import woowacourse.movie.movie.MovieMockData
+import woowacourse.movie.data.MovieMockData
 
 class HomeFragmentTest {
     @get:Rule
