+++ app/src/main/java/woowacourse/movie/fragment/movielist/HomeFragment.kt
@@ -12,7 +12,7 @@ import woowacourse.movie.R
 import woowacourse.movie.activity.moviedetail.MovieDetailActivity
 import woowacourse.movie.fragment.movielist.adapter.MovieRecyclerViewAdapter
 import woowacourse.movie.model.AdUIModel
-import woowacourse.movie.movie.MovieMockData
+import woowacourse.movie.data.MovieMockData
 import woowacourse.movie.movie.toPresentation
 
 class HomeFragment : Fragment(), HomeContract.View {
